+++ Firebase.js
@@ -3,7 +3,7 @@
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.8.0/firebase-app.js";
 import { getAnalytics } from "https://www.gstatic.com/firebasejs/10.8.0/firebase-analytics.js";
 import { getFirestore, collection, addDoc, query, where, getDocs } from "https://www.gstatic.com/firebasejs/10.8.0/firebase-firestore.js";
-import { consultar_cita } from "./consultar_cita.js";
+// import { consultar_cita } from "./consultar_cita.js";
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
@@ -24,4 +24,46 @@ const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
 
 let search = document.getElementById('obtainData');
-search.addEventListener('click',consultar_cita);
+search.addEventListener('click',async()=>{
+    const cc = document.getElementById('cedMostrar').value;
+    const q = query(collection(db,"REAL_DATA"),where("cc","==",cc));
+    try{
+        const querySnapshot = await getDocs(q);
+                    if (querySnapshot.empty){
+                console.log(`No hay registros con el CC ${cc}`);
+            } else {
+                querySnapshot.forEach((doc)=>{
+                    console.log(doc.id, ' => ', doc.data());
+                    let nombrePersona = doc.data().nombrePersona;
+                    let tipoAnimal = doc.data().tipoAnimal;
+                    let cc = doc.data().cc
+                    let raza = doc.data().raza;
+                    let fecha_cita = doc.data().fecha_cita;
+
+                    let campoPersona = document.createElement('th');
+                    let campoAnimal = document.createElement('th');
+                    let campoCC = document.createElement('th');
+                    let campoRaza = document.createElement('th');
+                    let campoFecha = document.createElement('th');
+
+                    campoPersona.textContent=nombrePersona;
+                    campoAnimal.textContent=tipoAnimal;
+                    campoCC.textContent=cc;
+                    campoRaza.textContent=raza;
+                    campoFecha.textContent=fecha_cita;
+                    let container = document.getElementById('tableBody');
+                    let tableCont = document.createElement('tr');
+                    tableCont.appendChild(campoPersona);
+                    tableCont.appendChild(campoCC);
+                    tableCont.appendChild(campoFecha);
+                    tableCont.appendChild(campoAnimal);
+                    tableCont.appendChild(campoRaza);
+                    container.appendChild(tableCont);
+                    console.log(nombrePersona,tipoAnimal,cc,raza,fecha_cita);
+                })
+            }
+    }
+    catch(error){
+        console.log('Error al obtener documentos: ', error)
+    }
+});
