+++ index.html
@@ -141,10 +141,11 @@
                     <table class="table is-striped" style="width: 100%;">
                         <thead>
                             <tr>
-                                <th title="Animal">Animal</th>
-                                <th title="Breed">Breed</th>
-                                <th title="Razon">Razon de Cita</th>
+                                <th title="Name">Owner Name</th>
+                                <th title="cc">ID</th>
                                 <th title="Date">Date</th>
+                                <th title="Animal">Animal</th>
+                                <th title="Raza">Raza</th>
                             </tr>
                         </thead>
                         <tbody id="tableBody">
